# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Taxi Dispatch",
      "version": "1.0",
      "description": "API pour la gestion des utilisateurs, des véhicules, des zones, et des trajets dans le système Taxi Dispatch."
    },
    "servers": [
      {
        "url": "http://localhost:8081/api"
      }
    ],
    "paths": {
      "/user/get_all": {
        "post": {
          "summary": "Obtenir tous les utilisateurs",
          "operationId": "getAllUsers",
          "responses": {
            "200": {
              "description": "Liste des utilisateurs",
              "content": {
                "application/json": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "/user/log_in": {
        "post": {
          "summary": "Connexion de l'utilisateur",
          "operationId": "logIn",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Réponse de connexion réussie",
              "content": {
                "application/json": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Identifiants incorrects"
            }
          }
        }
      },
      "/user/get_role": {
        "post": {
          "summary": "Obtenir le rôle d'un utilisateur",
          "operationId": "getUserRole",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    }
                  },
                  "required": ["user_id"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Rôle de l'utilisateur",
              "content": {
                "application/json": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/zone/get_all": {
        "post": {
          "summary": "Obtenir les zones",
          "operationId": "getAllZones",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    }
                  },
                  "required": ["user_id"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Liste des zones",
              "content": {
                "application/json": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Zone"
                  }
                }
              }
            }
          }
        }
      },
      "/vehicle/get_all": {
        "post": {
          "summary": "Obtenir tous les véhicules",
          "operationId": "getAllVehicles",
          "responses": {
            "200": {
              "description": "Liste des véhicules",
              "content": {
                "application/json": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          }
        }
      },
      "/driver/get_all": {
        "post": {
          "summary": "Obtenir tous les conducteurs",
          "operationId": "getAllDrivers",
          "responses": {
            "200": {
              "description": "Liste des conducteurs",
              "content": {
                "application/json": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              }
            }
          }
        }
      },
      "/trip/get_all": {
        "post": {
          "summary": "Obtenir tous les trajets",
          "operationId": "getAllTrips",
          "responses": {
            "200": {
              "description": "Liste des trajets",
              "content": {
                "application/json": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          }
        }
      },
      "/trip/get_orders": {
        "post": {
          "summary": "Obtenir toutes les commandes des trajets",
          "operationId": "getAllOrders",
          "responses": {
            "200": {
              "description": "Liste des commandes",
              "content": {
                "application/json": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "/trip/Accept": {
        "put": {
          "summary": "Accepter un voyage",
          "operationId": "acceptTrip",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trip_id": {
                      "type": "integer"
                    },
                    "driver_id": {
                      "type": "integer"
                    }
                  },
                  "required": ["trip_id", "driver_id"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Voyage accepté",
              "content": {
                "application/json": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "username": {
              "type": "string"
            },
            "role": {
              "type": "string"
            }
          }
        },
        "Zone": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            }
          }
        },
        "Vehicle": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "model": {
              "type": "string"
            }
          }
        },
        "Driver": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            }
          }
        },
        "Trip": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "start_time": {
              "type": "string",
              "format": "date-time"
            },
            "end_time": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Order": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "description": {
              "type": "string"
            }
          }
        }
      }
    }
  }
  